backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-celery-beat==2.5.0
celery==5.3.4
redis==5.0.0
gunicorn==21.0.0
psycopg2-binary==2.9.6
